refactor(media): segregate catalog repositories into role-interfaces
