+++ src/modules/media/infrastructure/persistence/repositories/movie_repository.py
@@ -15,10 +15,10 @@ from src.building_blocks.application.pagination import (
 )
 from src.building_blocks.domain.pagination import PaginatedResult, Pagination
 from src.modules.media.domain.entities import Movie
-from src.modules.media.domain.repositories import MovieRepository
-from src.modules.media.domain.repositories.movie_repository import (
+from src.modules.media.domain.repositories import (
     CreditsStatusRow,
     GenreRow,
+    MovieRepository,
     RemoteArtworkRow,
 )
 from src.modules.media.domain.value_objects import (
